Refactor implementations of dominant_swatches and swatches_with_theme

+++ tests/extract.rs
@@ -14,7 +14,7 @@ fn extract(#[case] path: &str, #[case] n: usize, #[case] expected: Vec<&str>) {
     let img = image::open(path).unwrap();
     let image_data = SimpleImageData::new(img.width(), img.height(), img.as_bytes()).unwrap();
 
-    let palette: Palette<f32> = Palette::extract(&image_data);
+    let palette: Palette<f64> = Palette::extract(&image_data);
     let swatches = palette.dominant_swatches(n);
     assert_eq!(swatches.len(), n);
 
@@ -30,7 +30,7 @@ fn extract_with_gmeans() {
     let img = image::open("./tests/images/aLMeYMZEJvk.png").unwrap();
     let image_data = SimpleImageData::new(img.width(), img.height(), img.as_bytes()).unwrap();
 
-    let palette: Palette<f32> = Palette::extract_with(&image_data, Algorithm::GMeans);
+    let palette: Palette<f64> = Palette::extract_with(&image_data, Algorithm::GMeans);
     let swatches = palette.dominant_swatches(4);
     assert_eq!(swatches.len(), 4);
 }
@@ -40,7 +40,7 @@ fn extract_with_dbscan() {
     let img = image::open("./tests/images/aLMeYMZEJvk.png").unwrap();
     let image_data = SimpleImageData::new(img.width(), img.height(), img.as_bytes()).unwrap();
 
-    let palette: Palette<f32> = Palette::extract_with(&image_data, Algorithm::DBSCAN);
+    let palette: Palette<f64> = Palette::extract_with(&image_data, Algorithm::DBSCAN);
     let swatches = palette.dominant_swatches(5);
     swatches.iter().for_each(|swatch| {
         println!(
@@ -60,7 +60,7 @@ fn extract_with_hdbscan() {
     let img = image::open("./tests/images/aLMeYMZEJvk.png").unwrap();
     let image_data = SimpleImageData::new(img.width(), img.height(), img.as_bytes()).unwrap();
 
-    let palette: Palette<f32> = Palette::extract_with(&image_data, Algorithm::HDBSCAN);
+    let palette: Palette<f64> = Palette::extract_with(&image_data, Algorithm::HDBSCAN);
     let swatches = palette.dominant_swatches(4);
     assert_eq!(swatches.len(), 4);
 }
